Add files via upload
A broken version that was available for download was salvaged as the creator (SuperiorFox) deleted the game in his browser version with his itch.io page. The reasons why the author deleted it is because of misunderstandings with people for details in the game content.

If anyone manages to make it work properly, please share their work for its preservation as well.